internal/fl_service/client_jetson/requirements.txt
flwr==1.7.0
flwr_datasets==0.5.0

numpy>=1.26,<2.0
pyyaml
Pillow>=10.0.0

pandas>=2.0,<3.0
scipy>=1.10,<2.0
scikit-learn>=1.3,<2.0

torch_geometric
geopy